Feat: Implement RWD

+++ app/components/ProtocolDeleteButton/ProtocolDeleteButton.tsx
@@ -6,21 +6,29 @@ import { useRouter } from "next/navigation";
 
 interface ProtocolDeleteButtonProps {
   id: string;
+  isInProtocol?: boolean;
 }
 
 export default function ProtocolDeleteButton({
   id,
+  isInProtocol = false,
 }: ProtocolDeleteButtonProps) {
   const router = useRouter();
   const onDelete = useGetDeleteHandler();
 
   const handleOnDelete = () => {
-    onDelete(id);
-    router.back();
+    if (window.confirm("Czy na pewno chcesz usunąć protokół?")) {
+      onDelete(id);
+      router.back();
+    }
   };
   return (
-    <button onClick={handleOnDelete}>
-      <DeleteIcon />
+    <button
+      onClick={handleOnDelete}
+      className={`${isInProtocol ? "flex items-center gap-2 uppercase hover:text-primary" : ""}`}
+    >
+      <DeleteIcon className="h-8 w-8 cursor-pointer hover:stroke-primary" />
+      {isInProtocol && "Usuń"}
     </button>
   );
 }
